backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20
redis==5.2.1
celery==5.4.0
stripe==11.2.0
twilio==9.3.7
emails==0.6
openpyxl==3.1.5
reportlab==4.2.5
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==33.1.0
factory-boy==3.3.1
